pyproject.toml
[tool.poetry]
name = "conjurer"
version = "0.1.0"
description = "libvirt API web frontend"
authors = ["Tyler Darnell"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
libvirt-python = "^6.10.0"
pytest = "^6.2.1"
coverage = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
